requirements_gradio.txt
# Core
gradio>=6.0.1
python-dotenv==1.0.0

# OpenAI APIs
openai>=2.0.0

# Audio Processing
pydub==0.25.1

# Image Processing
Pillow==10.1.0

# Utilities
requests>=2.32.0

# RAG / Vector Database
pinecone>=8.0.0
